Fix/entk script giannis

+++ src/predicting/predict_sealnet.py
@@ -89,7 +89,7 @@ def get_xy_locs(array, count, min_dist=3):
     return np.array([(x // cols, x % cols) for x in flat_order[:count]])
 
 
-def predict_patch(model, output_dir, test_dir, batch_size=2, input_size=299, threshold=0.5, num_workers=1):
+def predict_patch(model, output_dir, test_dir, batch_size=2, input_size=299, threshold=0.5, num_workers=1, remove_tiles=False):
     """
     Patch prediction function. Outputs shapefiles for counts and locations.
 
@@ -100,6 +100,7 @@ def predict_patch(model, output_dir, test_dir, batch_size=2, input_size=299, thr
     :param input_size: size of input images
     :param threshold: threshold for occupancy
     :param num_workers: number of workers on dataloader
+    :param remove_tiles: Remove the tiles folder from the filesystem.
     :return:
     """
 
@@ -176,16 +177,17 @@ def predict_patch(model, output_dir, test_dir, batch_size=2, input_size=299, thr
                                 'filenames': [ele for ele in fnames]})
 
     time_elapsed = time.time() - since
-    print('Testing complete in {}h {:.0f}m {:.0f}s'.format(
-        time_elapsed // 3600, time_elapsed // 60, time_elapsed % 60))
+    print('Testing complete in %dh %dm %ds' % (time_elapsed // 3600,
+                                               time_elapsed // 60,
+                                               time_elapsed % 60))
 
     # save shapefile for counts / classes
-    shapefile_path = '{}/predicted_shapefiles/{}/'.format(output_dir, os.path.basename(test_dir)[:-4])
+    shapefile_path = '%s/predicted_shapefiles/' % output_dir
     os.makedirs(shapefile_path)
 
     # load affine matrix
     affine_matrix = rasterio.Affine(*[ele for ele in pd.read_csv(
-        '{}/affine_matrix.csv'.format(test_dir))['transform']])
+        '%s/affine_matrix.csv' % (test_dir))['transform']])
 
     # create geopandas DataFrames to store counts per patch and seal locations
     output_shpfile = gpd.GeoDataFrame()
@@ -235,9 +237,10 @@ def predict_patch(model, output_dir, test_dir, batch_size=2, input_size=299, thr
             os.path.basename(output_dir)))
 
         # remove tiles
-        shutil.rmtree('{}/tiles'.format(test_dir))
+        if remove_tiles:
+            shutil.rmtree('{}/tiles'.format(test_dir))
 
-        print('    Total predicted in {}: '.format(os.path.basename(test_dir)), sum(pred_counts['predictions']))
+    print('Total predicted in %s: '% os.path.basename(test_dir), sum(pred_counts['predictions']))
 
 
 def main():
@@ -246,7 +249,7 @@ def main():
 
     # check for invalid inputs
     assert args.model_architecture in model_archs, "Invalid architecture -- see supported" \
-                                                   " architectures:  {}".format(list(model_archs.keys()))
+                                                   " architectures: %s" % list(model_archs.keys())
 
     assert args.hyperparameter_set in hyperparameters, "Hyperparameter combination is not defined in " \
                                                        "./utils/model_library.py"
@@ -271,7 +274,8 @@ def main():
     predict_patch(model=model, input_size=model_archs[args.model_architecture]['input_size'],
                   test_dir=args.test_dir, output_dir=args.output_dir,
                   batch_size=hyperparameters[args.hyperparameter_set]['batch_size_test'],
-                  num_workers=hyperparameters[args.hyperparameter_set]['num_workers_train'])
+                  num_workers=hyperparameters[args.hyperparameter_set]['num_workers_train'],
+                  remove_tiles=True)
 
 
 if __name__ == '__main__':
